+++ CHANGELOG.md
@@ -2411,3 +2411,4 @@
 - [2026-06-29 16:26] feat(test): add async test stream provider
 - [2026-06-29 16:35] test(storage): add unit tests for storage business logic
 - [2026-06-29 17:29] feat(theme): implement shimmer loading skeleton in theme
+- [2026-06-29 17:55] refactor(network): extract reusable atomic widgets from network